Agent: Move PortScanData to infection_monkey/dataclasses/ to avoid circular import
