+++ monkey/infection_monkey/dataclasses/__init__.py
@@ -1,0 +1,1 @@
+from .port_scan_data import PortScanData
