+++ monkey/infection_monkey/i_puppet/i_fingerprinter.py
@@ -1,9 +1,10 @@
 from abc import abstractmethod
 from typing import Dict
 
+from infection_monkey.dataclasses import PortScanData
 from infection_monkey.i_puppet import PingScanData
 
-from . import FingerprintData, PortScanData
+from . import FingerprintData
 
 
 class IFingerprinter:
