+++ monkey/infection_monkey/network_scanning/smb_fingerprinter.py
@@ -7,7 +7,8 @@ from odict import odict
 
 from common import OperatingSystem
 from common.types import PortStatus
-from infection_monkey.i_puppet import FingerprintData, IFingerprinter, PingScanData, PortScanData
+from infection_monkey.dataclasses import PortScanData
+from infection_monkey.i_puppet import FingerprintData, IFingerprinter, PingScanData
 
 DISPLAY_NAME = "SMB"
 SMB_PORT = 445
